Move runparts to subp.
runparts (run a directory of scripts) seems to fit well in subp
module.  The request to move it there was raised in #416.

Replace use of logexc with LOG.debug as logexc comes from util.
